Error check finished, need to organization
