+++ maps/easy.cub
@@ -14,7 +14,7 @@ C 225,100,0
 111111111111 111111111111
 1000000000001000000000001
 1000000000000000000000001
-1000000000000000000000001
+1000000000S00000000000001
 1000000000000000000000011
 1000000000000000000000001
 1111111111111111111111111
