+++ maps/subject.cub
@@ -9,7 +9,7 @@ C 225,30,0
 1111111111111111111111111
 1000000000110000000000001
 1011000001110000000000001
-1001000000000000000000001
+10010000000000000000000011
 111111111011000001110000000000001
 100000000011000001110111111111111
 11110111111111011100000010001
