+++ maps/wrong_char_in_map.cub
@@ -8,7 +8,7 @@ C 255, 255, 255
 
 111111
 100001111
-10W 01111
+10W201111
 100000001
 111011111
   101
